+++ vamp-plugins/plugins/BeatTrack.h
@@ -57,9 +57,9 @@ protected:
     int m_dfType;
 
     // MEPD new protected parameters to allow the user to control these advanced parameters of the beat tracker
-    double m_alpha;
-    double m_tightness;
-    double m_inputtempo;
+    float m_alpha;
+    float m_tightness;
+    float m_inputtempo;
     bool m_constraintempo;
 
     bool m_whiten;
